Refactored the FetchMovieTask and all of it's components. Added addSortOrder to the FetchMovieTask and a bulkInsert.
Changed getMoviesDataFromJson to match Sunshine example and use the db for populating the adapter, instead of doing so externally.
Changed onPostExecute to use a more efficient for loop.

Exercise 4B Video 23 (Last) - BulkInserts with ContentProvider

The bulkInsert does not function as expected. The ContentResolver.bulkInsert returns the correct number of inserts, however the cursor itself return incorrect values (not updated).