Refactored the FetchMovieTask and all of it's components. Added addSortOrder to the FetchMovieTask and a bulkInsert.
Changed getMoviesDataFromJson to match Sunshine example and use the db for populating the adapter, instead of doing so externally.
Changed onPostExecute to use a more efficient for loop.

Exercise 4B Video 23 (Last) - BulkInserts with ContentProvider

The problem was inside the DBHelper- in particular, the statement where we assure just one movie record from each sort order.
 Added a second row from the movie record to be compared.
All is working correctly.