Refactored the functions inside MoviesContrac: - createMovieWithDate - creaMovieWIthStartDate with something that is actually useful. Now the create uri by querying the movie ID. Also change the conditions inside the MovidDBHelper to check if the movie id is identical instead of the movie name, as the movies can actually have the same name.
Changed the Details activity to take information from the database and display it.

Changed the onClick functionality.
It is declared inside the ViewHolder class.
To get the correct item from the cursor, the cursor is moved to the position from the adapter that is touched.
Then the information is acquired from the said position

Exercise 4C part 11 - Make Details View Functional.