Modified (New): - Added touch_selector for phones and tablets (sw600dp) - The background if for the main theme. - Now the selected item is properly highlighted for the item. - Added a function to get and reset the mSelectedItem position. - In MovieGridFragment the position of the selected item is reset, when the sort order changes.
Modified (Old):
-Main Activity
-MovieGridFragment
-GridAdapter (onClick)
-DetailActivity
-DetailFragment

As the app now supports 2-pane view it has to have a modified behaviour when clicking on an item.
When an Item is clicked, instead of always launching a new intent to the DetailActivity, we notify the CallBack instead.
We get the context and cast it to a CallBack object. In this solution the onClick function is inside the RecyclerView adapter- GridAdapter.

Inside the MainActivity we implement the CallBack interface.
Inside the onItemSelected method, we define how the app will react if we are in 2-pane or single-pane mode.
In case it is 2-pane, we pass the URI to the DetailFragment and replace the container. In single-pane mode, we create an Intent as usual.

In DetailActivity, the onCreate method is modified to read the incoming URI from the intent. We take the URI and set it as arguments for the DetailFragment. Then we dynamically add the fragment to the container.

In the DetailFragment, in onCreateView method, the fragment reads the arguments with which it was initialized.
Inside DetailFragment, the onCreateLoader now uses mUri instead of relying on the incoming Intent.

Inside the SettingsActivity, a small patch was added to prevent from a bug.

TODO: optimize the target image size
Lesson 5- Video 29 "Handle List Item Click"